requirements.txt
beautifulsoup4
requests
python-dotenv
telegraph